safari-534.53-branch/LayoutTests/http/tests/media/video-cross-site-expected.txt

Video with cross-site reference should not load.

EXPECTED (video.error == 'null') OK
EVENT(loadstart)
EVENT(error)
SUCCESS: denied cross-site video load. OK

END OF TEST
